refactor: RecommendedArtworks section
